Changed utils to return a float64

+++ utils/utils.go
@@ -4,14 +4,14 @@ import (
 	"math"
 )
 
-func BytesToKB(b uint64) uint64 {
-	return uint64((float64(b) / math.Pow10(3)))
+func BytesToKB(b uint64) float64 {
+	return float64(b) / math.Pow10(3)
 }
 
-func BytesToMB(b uint64) uint64 {
-	return uint64((float64(b) / math.Pow10(6)))
+func BytesToMB(b uint64) float64 {
+	return float64(b) / math.Pow10(6)
 }
 
-func BytesToGB(b uint64) uint64 {
-	return uint64((float64(b) / math.Pow10(9)))
+func BytesToGB(b uint64) float64 {
+	return float64(b) / math.Pow10(9)
 }
